Fixed issue where QuadraticSlider used ints instead of floats for no good reason

+++ Scripts/Core/DrawerAttributes/QuadraticSliderAttribute.cs
@@ -2,11 +2,11 @@
 {
     public class QuadraticSliderAttribute : DrawerAttribute
     {
-        public int Power { get; private set; }
-        public int Min { get; private set; }
-        public int Max { get; private set; }
+        public float Power { get; private set; }
+        public float Min { get; private set; }
+        public float Max { get; private set; }
 
-        public QuadraticSliderAttribute(int min, int max, int power)
+        public QuadraticSliderAttribute(float min, float max, float power)
         {
             Min = min;
             Max = max;
